pyproject.toml
[project]
name = "redmine-mcp-server"
version = "0.12.1"
description = "Production-ready MCP server for Redmine with security, pagination, and enterprise features"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["redmine", "mcp", "model-context-protocol", "project-management", "server"]
authors = [
    {name = "Kevin Tan", email = "jingzheng.tan@gmail.com"}
]
maintainers = [
    {name = "Kevin Tan", email = "jingzheng.tan@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Groupware",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi[standard]>=0.120.0",
    "httpx>=0.27.0",
    "mcp[cli]>=1.25.0,<2",
    "python-redmine>=2.5.0",
    "uvicorn>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/jztan/redmine-mcp-server"
Repository = "https://github.com/jztan/redmine-mcp-server.git"
Issues = "https://github.com/jztan/redmine-mcp-server/issues"
Changelog = "https://github.com/jztan/redmine-mcp-server/blob/master/CHANGELOG.md"

[project.scripts]
redmine-mcp-server = "redmine_mcp_server.main:main"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "httpx>=0.28.1",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "httpx>=0.28.1",
    "flake8>=6.0.0",
    "black>=23.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "pragma: no cover",
]